crawler/requirements.txt
# Web scraping and HTML parsing
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
selenium>=4.15.0
playwright>=1.40.0

# PDF processing
PyMuPDF>=1.23.0

# Vertex AI and LangChain
langchain-google-vertexai>=1.0.0
langchain-core>=0.3.0
langchain-community>=0.3.0
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
vertexai>=1.38.0  # For Gemini model (natural answer generation)
# Note: tensorflow-text is only needed if using VectorSearchVectorStoreGCS
# Use upload_vectors_direct.py instead to avoid TensorFlow dependency
# tensorflow-text>=2.15.0

# Utilities
python-dotenv>=1.0.0
flask>=3.0.0
